Update with comments from review Check all parameters Rename some types to not use underscores Removed posix files from test build Default stack size and priority fixes Update taskpool timer arm to 32-bit safe call

+++ lib/include/platform/iot_threads.h
@@ -334,7 +334,7 @@ bool IotSemaphore_TryWait( IotSemaphore_t * const pSemaphore );
  */
 /* @[declare_platform_threads_semaphoretimedwait] */
 bool IotSemaphore_TimedWait( IotSemaphore_t * const pSemaphore,
-                             uint64_t timeoutMs );
+                             uint32_t timeoutMs );
 /* @[declare_platform_threads_semaphoretimedwait] */
 
 /**
